Support controlling red LED strips on side of FS67xx devices
It's those red strips on the left and right side of the power button.
They're controlled by three GPIOs that are exposed as the new LEDs
red:side_inner, red:side_mid and red:side_outer

One of them was previously controlled by power:front_panel, but I
changed it so it's more consistent with the other two LEDs (LED pairs?)
of the strips

Thanks to @brunomatic for figuring out the GPIOs of these LEDs!